desain admin dan user

+++ app/Http/Controllers/Menu/DataGuruController.php
@@ -5,20 +5,44 @@ namespace App\Http\Controllers\Menu;
 use App\Http\Controllers\Controller;
 use App\Models\DataGuru;
 use Illuminate\Http\Request;
-use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Carbon;
 use Illuminate\Support\Str;
+use Yajra\DataTables\Facades\DataTables;
 
 class DataGuruController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $DataGuru = DataGuru::all();
-        return view('menu/dataguru/list', compact('DataGuru'));
+        if ($request->ajax()) {
+            $model = DataGuru::query()->select('id', 'nama', 'file_foto');
+            if ($request->has('order') && !empty($request->input('order'))) {
+                $order = $request->input('order')[0];
+                $columnIndex = $order['column'];
+                $columnName = $request->input('columns')[$columnIndex]['data'];
+                $columnDirection = $order['dir'];
+                $model->orderBy($columnName, $columnDirection);
+            } else {
+                $model->latest();
+            }
+            return DataTables::eloquent($model)
+                ->addIndexColumn()
+                ->editColumn('file_foto', function ($row) {
+                    $imageUrl = asset('img/' . $row->file_foto);
+                    return "<img src='{$imageUrl}' class='w-64-px h-64-px radius-8 object-fit-cover' alt='Avatar'>";
+                })
+                ->addColumn('action', function ($row) {
+                    return '<a class="w-32-px h-32-px bg-primary-light text-primary-600 rounded-circle d-inline-flex align-items-center justify-content-center" href="' . route('kelola-data-guru.show', $row->id) . '"> <iconify-icon icon="iconamoon:eye-light"></iconify-icon> </a> <a class="w-32-px h-32-px bg-success-focus text-success-main rounded-circle d-inline-flex align-items-center justify-content-center" href="' . route('kelola-data-guru.edit', $row->id) . '"> <iconify-icon icon="lucide:edit"></iconify-icon> </a> <a class="w-32-px h-32-px bg-danger-focus text-danger-main rounded-circle d-inline-flex align-items-center justify-content-center" style="cursor:pointer" onclick="hapus(\'' . $row->id . '\')"> <iconify-icon icon="mingcute:delete-2-line"></iconify-icon> </a>';
+                })
+                ->rawColumns(['file_foto', 'action'])
+                ->removeColumn('id')
+                ->toJson(true);
+        }
+        return view('menu/dataguru/list');
     }
 
     /**
@@ -92,7 +116,7 @@ class DataGuruController extends Controller
     {
         $rules = [
             'nama-guru' => 'required|string|max:255',
-            'nip-guru' => 'required|digits:18|unique:data_guru,nip,' . $id,
+            'nip-guru' => 'required|digits:18|unique:tb_data_guru,nip,' . $id,
             'bidang-guru' => 'required|string',
             'jabatan-guru' => 'required|string',
             'foto-guru' => 'nullable|image|mimes:jpeg,jpg,png|max:3072',
@@ -124,19 +148,32 @@ class DataGuruController extends Controller
         $getDataGuru->update($data);
 
         return redirect()->route('kelola-data-guru.index')->with(['message' => 'sukses mengubah data.', 'isActive' => true, 'hasError' => false]);
-
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Request $request, string $id)
     {
-        $DataGuru = DataGuru::findOrFail($id);
-        if (!empty($DataGuru->file_foto)) {
-            Storage::delete($DataGuru->file_foto);
+        if ($request->ajax()) {
+            try {
+                $DataGuru = DataGuru::findOrFail($id);
+                if (!empty($DataGuru->file_foto)) {
+                    Storage::delete($DataGuru->file_foto);
+                }
+                $DataGuru->delete();
+
+                return response([
+                    'status' => 'success',
+                    'data' => 'sukses menghapus data.'
+                ]);
+            } catch (\Exception $e) {
+                return response([
+                    'status' => 'error',
+                    'data' => $e->getMessage()
+                ], 500);
+            }
         }
-        $DataGuru->delete();
-        return redirect()->route('kelola-data-guru.index')->with('message', 'sukses menghapus data.');
+        return redirect()->back();
     }
 }
